Implement Test Pyramid
Separates tests into a Test Pyramid as proposed by
Martin Fowler. Currently only has unit and component
tests. Integration is not ready as there is no database
beyond SQLite. Contract is not ready as there is no
deployed HTTP functions. End-to-end is for the same
reason along with a lack of a web front end.