Add database integration tests
Adds integration tests for the ORM to the database. These are currently
skipped until a real database is in place. In the mean time, they can be
used with a local Postgres Docker container.

Creates a database class hierarchy which is used to store and create
various connection strings.

Moves the config file to the root of the app.

Chnages the manage script to run integrated and non-integrated tests
separately.

Moves fixture functionality into helper functions.

Fixes the config unit tests as they were actually creating a real
database connection instead of just testing config.