Replace Flask Script with Click
As Click is now abandoned and the maintainers recommend Click, Click it
shall be.

Moves the functionality of manage.py to api_skeleton.py and then adds
new features, then removes manage.py.

Creates separate integrated and non-integrated config classes rather
than relying on environment variables. This is because integration is
now set as a command line argument to Click. We want to avoid a mismatch
where one is set and the other isn't.

Updates the horrible and soon to be removed run.bash to suit this brave
new world.