Replace runner script
Removes the horrible BASH script and replaces it with a shiny new Python
script. It does everything the old script does and also adds
python-dotenv to load environment variables during execution rather than
before hand in the BASH script.

Renames the .conf files to .env to be more consistent with how the rest
of the world operates. Also moves the common environment variables to a
common file.