Add Pylint runner and badge
Adds lint command to api_skeleton.py which programmatically invokes
pylint. It then uses anybadge to create an svg file of the score.

Updates the CI runner to use the renamed tools-requirements.txt file and
the updated runner.py to generate the svg file. Finally, this file is
uploaded.

Changes made in the GitLab UI to find the svg and display it.