Start local test dependencies when running in local mode
When the environment variable LOCAL is set to true, the runner will
start the Docker Postgres database container, upgrade the DDL and start
a Flask app. This allows for local debugging of an "integrated"
database and "deployed" HTTP Flask endpoint.

The Flask app is in separate daemon thread because otherwise it would
begin and never return to the script as it is a foreground process.