Dump Swagger to Postman collection and environment
Uses the as_postman() method from Flask-RESTPlus to create a Postman
collection. Takes this collection and adds configuration to it such as
headers, bodies, Javascript snippets and template variables. Then dumps
the JSON to file.

Creates the values for the template variables and also dumps these to a
separate JSON file for an environment.

These files allow for a fully automated creation of a Postman workspace.
The API has templated variables where needed and these are either hard-
coded or set dynamically. The dynamic values are set using the
Javascript snippets as Postman tests; these are run at the end of the
request. This way you set things such as authorisation tokens or the
public ID of the user in the correct part of the development workflow.
You also have the correct request body schemas and the variables to use
them.

In summary, you can pass a complete development environment to a fellow
developer and they will think you are a boss. Which is nice. You're
welcome.