Remove deprecated and unused fields
Replaces the 'X-API-KEY' header with 'Authorization' (sic/fully sick).
Unfortunately Flask-RESTPlus only supports OpenAPI 2.0 so we're stuck
with a bit of a fiddle in the Swagger docs. Basically pretending that a
JWT is an API key. Sigh. Software development is pragmatic hacks.

Removes the api.parser() method from /auth/logout because setting the
above overrides it for all endpoints and fairly soon they will actually
be protected by the bearer token.

Because of the above instead of setting the header in the Postman script
for particular routes, we just apply an 'auth' key to the root of the
Swagger/Postman JSON dump.

Renames a bunch of functions so they are both sensibly named and also
marked as private. So keep your nose out.