Make marshalling consistent
Creates a base response class to use in marshalling. This restructures
the response payload to use the JSend specification. This is a very
simple and lightweight spec for returning JSON over REST.

Creates a helper method for deserialising a User object. It is nicer to
use marshalling for this but so far I can't figure out how to override
the data field in the new response DTO. Future improvement.

Removes 200 responses decorators as they can be set in the marshall_with
decorator instead.

Moves overlooked response strings to the responses module. Cleans up the
same module.

Fixes all the broken tests and Postman stuff from this madness.