Create response helper
Creates a responder function to consolidate all response construction.
The function takes the HTTP status code and optionally a data
dictionary. Codes 499 and under use the dictionary and add it the 'data'
element in the HTTP response. Codes 399 and under send a success status.
400 to 499 send a fail status. Codes 500 and above do not have a data
element; they have a 'message' element. This is to meet the JSend spec.

This replaces a few wayward errors which were returning an error payload
directly instead of raising to the top-level handler. Success payloads
continue to return directly from the function but they now use the
responder function.

A call to abort has been removed and been replaced with a raise. Future
designs will consider whether to use abort or raise; but this change is
aimed at making all errors be handled the same way.

Adds a base level config setting to remove the automatic addition by
Flask of a 'message' attribute to the response. This attribute conflicts
with the one for JSpec and also is not able to be controlled manually.
Kill that noise.

Consolidates the error handlers into a single top level one because when
all the various errors were mapped out, they were all handled the same
way. A future design may change this in a case where errors are handled
differently. Also note that this error handler catches the base
Exception type. Using this decorator without an exception is supposed to
use the default exception handler which I would expect would catch all
unhandled errors; but it doesn't seem to. Insert shrug here.