Add internationalisation with Flask-Babel
Uses Flask-Babel to create a base language (English) and then create a
demonstration translated language (Australian English). All base strings
are in their own base module and using the power of gettext, these are
marked for translation. A local selection is added to the app init
module which chooses the best language match based on the request header
Accept-Language.

Now we speak Strine, mate.