Fix bugs in internationalisation
Replaces a hyphen with an underscore because that is actually a valid
separator. Funnily enough the invalid separator crashed Babel.

Fixes a bug where lazy loaded strings can't be JSON serialised. In the
app context this is fixed by overriding the JSONEncoder class to check
if it is an instance of speaklater._LazyString. If so it uses
flask._compat.text_type to infer that it is indeed a string which
funnily enough is JSON serialisable.

Is Flask RESTPlus where we don't have an app context, we need to
manually use the above procedure in the decorator. Will try to think of
a less lame solution.