Fix session cookie domain warning
When Flask is being served from an IP address, it emits a warning:
    UserWarning: The session cookie domain is an IP address. This may
    not work as intended in some browsers. Add an entry to your hosts
    file, for example "localhost.localdomain", and use that instead.

We would never serve this from an IP address in reality so to get rid of
this annoying warning locally we take that advice.

Updates the DEFAULT_LOCAL_SERVER constant in the config file as this is
the default for SERVER_NAME. Also updates this in the Postman config.

You will need to add this line to your /etc/host file:
127.0.0.1   localhost.localdomain

Or don't. I have and now I don't have that warning. It's your call.