Add authorisation
Extends the use of flask-jwt-simple by discovering the user from the JWT
they send. Uses this to restrict non-admin users from viewing anyone but
themselves.

This necessitated a number of tests to be fixed. Which led me down the
rabbit hole of interdependent fixtures. Refactored the rest of the test
suite to greatly simplify the use of fixtures, helpers and factories.
The code is much easier to read and debug now.