Add email confirmation for new users
When a user registers they must now confirm their email address before
accessing the API. This is achieved using Flask-Mail.

Creates an email client to send real email in PROD and dumps to a JSON
file in TEST. In DEV the API sends to a "real" mail server
transparently. That "real" server is a subclassed SMTPD DebuggingServer
which is overridden to similarly dump the data to a JSON file. Running
this DEV server has been added to the Click manager.

The email uses HTML and plaintext for the body.

The /login endpoint prevents unconfirmed emails from logging in.

JWTs are not returned from POSTing a new user any longer. This is now
only from /login when email confirmed.

Updates tests and Postman for new situation.

Updates database schema using Alembic.