Add token blacklisted decorator
Fixes a bug where blacklisted tokens are not checked. This allowed
viewing users after the user themselves had logged out of the
application. The check is in a decorator added to any route with the
jwt_required decorator. This decorator checks the header format and then
whether the token has been blacklisted.