Add password reset via confirmed email
Using a similar approach to the feature where users have to confirm
their email address before accessing the API, they can now reset their
password via email.

Two new endpoints receive the request and the token that the user
receives. The token is a timed JWT serialised using a different salt to
the confirmation salt.

The email client has been updated to have a generic function for
creating plaintext and HTML bodies.

Test have been added to cover the new scenario. As with the email
confirmation feature, the TEST environment intercepts the email sending
process and instead dumps the data to a JSON file for the suite to use.
And in DEV the sub-classed debugging SMTP server also dumps to JSON.

Postman has been updated to cater for the path variables reset_token and
confirmation_token.