Allow a user to change their email address.
Adds a new base DTO which contains only an email address. This allows
for further inheritance to later DTOs.

Updates the helper to check if the endpoint denies incorrect or
unauthorised tokens. Now data can be passed to this function.

Pushes some JSON dumping into the helper function instead of making
the callers do it.