Make endpoint naming consistent
As the endpoints developed over time they became inconsistent. This
updates so similar routes have the same root URL. This meant the
tests had to be updated for URLs and for where they are called from.

It also surfaced issues where methods/functions were in inconsistent
files, along with some methods really needing to functions. These have
been moved and a new common file has been created. It's not the nicest
but it solves what would be a resulting circular import error. Would be
nice to make this nicer. Nice.