Load environment variables from file
Updates the manager script to load environment variables from files,
on-the-fly. Certain variables are required to create the Flask app. For
example, if secrets aren't set, a ValueError is raised. This removes the
need to source the environment variables into the shell before starting.

The previous design of the manager imported the app when the module was
loaded. The new design imports the app only when required and after the
environment variables have been loaded by Python. The manager script
also knows which set of environment variables need to be loaded based on
the developer's command line arguments. This makes for a lower mental
overhead for the developer switching between various manager functions
and environments.

The runner now does not support launching the run command as loading the
environment variables in the manager instead makes this redundant.

The runner script and conftest now do not rely on the manager to create
a Flask app. This circular dependency technically worked because one
calling the other was done in a separate shell rather than in the same
Python process but it was pretty icky.

Renames some environment variables files for consistency and moves some
runner functions into their own functions for clarity.